Elim. redundancy by refactoring input components out of page definitions
