+++ src/components/index.js
@@ -1,4 +1,5 @@
 import Footer from './footer/Footer';
 import Navbar from './navbar/Navbar';
+import { TextInput, EmailInput, PasswordInput, PhoneInput } from './input';
 
-export { Footer, Navbar };
+export { Footer, Navbar, TextInput, EmailInput, PasswordInput, PhoneInput };
