+++ src/components/input/index.js
@@ -1,0 +1,147 @@
+import './input.scss';
+import { useState } from 'react';
+import { AiOutlineEye, AiOutlineEyeInvisible } from 'react-icons/ai';
+import { PatternFormat } from 'react-number-format';
+import Proptypes from 'prop-types';
+import { useActiveInput } from '../../utils';
+
+// TextInput: Utility component for handling text input
+export function TextInput({ labelText = 'Label', value, setter }) {
+  const { active, handleToggleActiveLabel } = useActiveInput();
+  return (
+    <div className={`input ${active ? 'active' : ''}`}>
+      <input
+        type="text"
+        id="text"
+        className="stretch"
+        value={value}
+        onChange={(evt) => {
+          setter(evt.target.value);
+          handleToggleActiveLabel(evt);
+        }}
+      />
+      <label htmlFor="text">{labelText}</label>
+    </div>
+  );
+}
+
+TextInput.propTypes = {
+  labelText: Proptypes.string.isRequired,
+  value: Proptypes.string.isRequired,
+  setter: Proptypes.func.isRequired,
+};
+
+// PasswordInput: Utility component for handling password input
+export function PasswordInput({ labelText = 'Label', value, setter }) {
+  const [isVisible, setIsVisible] = useState(false);
+  const { active, handleToggleActiveLabel } = useActiveInput();
+
+  const handleToggleVisibility = (evt) => {
+    evt.preventDefault();
+    setIsVisible(!isVisible);
+  };
+
+  return (
+    <div className={`input ${active ? 'active' : ''}`}>
+      <input
+        type={`${isVisible ? 'text' : 'password'}`}
+        id="password"
+        value={value}
+        onChange={(evt) => {
+          setter(evt.target.value);
+          handleToggleActiveLabel(evt);
+        }}
+      />
+      <label htmlFor="password">{labelText}</label>
+
+      {isVisible ? (
+        <AiOutlineEyeInvisible
+          className="toggle-visibility"
+          onClick={handleToggleVisibility}
+        />
+      ) : (
+        <AiOutlineEye
+          className="toggle-visibility"
+          onClick={handleToggleVisibility}
+        />
+      )}
+    </div>
+  );
+}
+
+PasswordInput.propTypes = {
+  labelText: Proptypes.string.isRequired,
+  value: Proptypes.string.isRequired,
+  setter: Proptypes.func.isRequired,
+};
+
+// EmailInput: Utility component for handling email input along with suggestions
+export function EmailInput({ labelText = 'Label', value, setter }) {
+  const { active, handleToggleActiveLabel } = useActiveInput();
+  return (
+    <div className={`input ${active ? 'active' : ''}`}>
+      <input
+        type="email"
+        id="email"
+        className="stretch"
+        value={value}
+        onChange={(evt) => {
+          setter(evt.target.value);
+          handleToggleActiveLabel(evt);
+        }}
+      />
+      <label htmlFor="email">{labelText}</label>
+    </div>
+  );
+}
+
+EmailInput.propTypes = {
+  labelText: Proptypes.string.isRequired,
+  value: Proptypes.string.isRequired,
+  setter: Proptypes.func.isRequired,
+};
+
+// PhoneInput: Utility component for handling phone input along with autoformatting
+export function PhoneInput({ labelText = 'Label', value, setter }) {
+  const { active, handleToggleActiveLabel } = useActiveInput();
+
+  // TODO: We may need to reformat how the phone number is actually stored in state
+  // Depending on the needs of the backend API.
+  const handlePhoneNumberUpdate = (value) => {
+    setter(value);
+  };
+
+  return (
+    <div className={`input ${active ? 'active' : ''}`}>
+      {/**I used PatternFormat in the code below because it gave me a
+      simpler way to implement automatic formatting for the phone number
+      as the user types.*/}
+      <PatternFormat
+        type="tel"
+        id="phone"
+        className="stretch"
+        value={value}
+        onChange={(evt) => {
+          console.log(value);
+          handlePhoneNumberUpdate(evt.target.value);
+          handleToggleActiveLabel(evt);
+        }}
+        format="+234 - #### - ### - ###"
+        mask="#"
+      />
+      <label
+        htmlFor="phone"
+        className="toggle-phone-number"
+        data-alt-name="Phone Number"
+      >
+        {labelText}
+      </label>
+    </div>
+  );
+}
+
+PhoneInput.propTypes = {
+  labelText: Proptypes.string.isRequired,
+  value: Proptypes.string.isRequired,
+  setter: Proptypes.func.isRequired,
+};
