+++ src/pages/login/Login.jsx
@@ -2,93 +2,7 @@ import './login.scss';
 import logo from '../../assets/small-logo.png';
 import { useState } from 'react';
 import { Link, useLocation, useNavigate } from 'react-router-dom';
-import { AiOutlineEye, AiOutlineEyeInvisible } from 'react-icons/ai';
-import Proptypes from 'prop-types';
-
-// Utility Hook: used to handle active state for inputs
-function useActiveInput() {
-  const [active, setActive] = useState(false);
-
-  const handleToggleActiveLabel = (evt) => {
-    if (evt.target.value === '') {
-      setActive(false);
-    } else {
-      setActive(true);
-    }
-  };
-
-  return { active, handleToggleActiveLabel };
-}
-
-// EmailInput: Utility component for handling email input along with suggestions
-function EmailInput({ labelText = 'Label', value, setter }) {
-  const { active, handleToggleActiveLabel } = useActiveInput();
-  return (
-    <div className={`input ${active ? 'active' : ''}`}>
-      <input
-        type="email"
-        id="email"
-        className="stretch"
-        value={value}
-        onChange={(evt) => {
-          setter(evt.target.value);
-          handleToggleActiveLabel(evt);
-        }}
-      />
-      <label htmlFor="email">{labelText}</label>
-    </div>
-  );
-}
-
-EmailInput.propTypes = {
-  labelText: Proptypes.string.isRequired,
-  value: Proptypes.string.isRequired,
-  setter: Proptypes.func.isRequired,
-};
-
-// PasswordInput: Utility component for handling password input
-function PasswordInput({ labelText = 'Label', value, setter }) {
-  const [isVisible, setIsVisible] = useState(false);
-  const { active, handleToggleActiveLabel } = useActiveInput();
-
-  const handleToggleVisibility = (evt) => {
-    evt.preventDefault();
-    setIsVisible(!isVisible);
-  };
-
-  return (
-    <div className={`input ${active ? 'active' : ''}`}>
-      <input
-        type={`${isVisible ? 'text' : 'password'}`}
-        id="password"
-        value={value}
-        onChange={(evt) => {
-          setter(evt.target.value);
-          handleToggleActiveLabel(evt);
-        }}
-      />
-      <label htmlFor="password">{labelText}</label>
-
-      {isVisible ? (
-        <AiOutlineEyeInvisible
-          className="toggle-visibility"
-          onClick={handleToggleVisibility}
-        />
-      ) : (
-        <AiOutlineEye
-          className="toggle-visibility"
-          onClick={handleToggleVisibility}
-        />
-      )}
-    </div>
-  );
-}
-
-PasswordInput.propTypes = {
-  labelText: Proptypes.string.isRequired,
-  value: Proptypes.string.isRequired,
-  setter: Proptypes.func.isRequired,
-};
+import { EmailInput, PasswordInput } from '../../components';
 
 export default function Login() {
   const location = useLocation();
