+++ src/utils/index.js
@@ -1,0 +1,3 @@
+import useActiveInput from './hooks/useactiveinput';
+
+export { useActiveInput };
